requirements.txt
streamlit
google-genai
fpdf
Pillow
pandas